+++ darwin/kinetic_player/Sources/kinetic_player/KineticPlayerPlugin.swift
@@ -1,10 +1,20 @@
+#if os(iOS)
 import Flutter
 import UIKit
+#elseif os(macOS)
+import AppKit
+import FlutterMacOS
+#endif
 
 public class KineticPlayerPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
+#if os(iOS)
+    let messenger = registrar.messenger()
+#elseif os(macOS)
+    let messenger = registrar.messenger
+#endif
     registrar.register(
-      SgVideoViewFactory(messenger: registrar.messenger()),
+      SgVideoViewFactory(messenger: messenger),
       withId: PlayerConstants.sgViewType
     )
   }