+++ macos/scripts/generate_package_swift.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-PLUGIN_ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")/../.." && pwd)"
-exec bash "${PLUGIN_ROOT}/darwin/scripts/sgplayer/generate_package_swift.sh" macos "$@"